working on load file
